backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
sqlalchemy==2.0.32
psycopg[binary]>=3.1
python-dotenv==1.0.1
pyjwt==2.8.0
pdfminer.six>=20231228